Convert wrap-jsonp.browser.js
