--- tests/browser/wrappings/wrap-jsonp.spec.js
@@ -1,2 +1,0 @@
-let matcher = require('../../../tools/jil/util/browser-matcher')
-module.exports = matcher.withFeature('addEventListener')
